automation/Tenjin/requirements.txt
selenium==4.9.0
webdriver-manager==3.8.6
python-dotenv==0.21.1
pandas==1.3.5
openpyxl==3.1.2